Rename ComplexDataStructure -> ComplexTestDataStructure
