+++ hazelcast/src/test/java/com/hazelcast/query/impl/extractor/specification/ExtractionInPortableSpecTest.java
@@ -38,12 +38,12 @@ import static com.hazelcast.query.impl.extractor.AbstractExtractionSpecification
 import static com.hazelcast.query.impl.extractor.AbstractExtractionSpecification.Index.ORDERED;
 import static com.hazelcast.query.impl.extractor.AbstractExtractionSpecification.Index.UNORDERED;
 import static com.hazelcast.query.impl.extractor.AbstractExtractionSpecification.Multivalue.PORTABLE;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.Finger;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.Person;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.finger;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.limb;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.person;
-import static com.hazelcast.query.impl.extractor.specification.ComplexDataStructure.tattoos;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.Finger;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.Person;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.finger;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.limb;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.person;
+import static com.hazelcast.query.impl.extractor.specification.ComplexTestDataStructure.tattoos;
 import static java.util.Arrays.asList;
 
 /**
@@ -82,7 +82,7 @@ public class ExtractionInPortableSpecTest extends AbstractExtractionTest {
         return new Configurator() {
             @Override
             public void doWithConfig(Config config, Multivalue mv) {
-                config.getSerializationConfig().addPortableFactory(ComplexDataStructure.PersonPortableFactory.ID, new ComplexDataStructure.PersonPortableFactory());
+                config.getSerializationConfig().addPortableFactory(ComplexTestDataStructure.PersonPortableFactory.ID, new ComplexTestDataStructure.PersonPortableFactory());
             }
         };
     }
